requirements.txt
matplotlib==3.0.2
numpy==1.17.2
Pillow==5.3.0
pyparsing==2.4.2
torch==1.2.0
torchvision==0.4.0
tqdm==4.35.0
